Update TileSet.xml for a clearer description on the bitmask data
so, I just spent a lot of time figuring out how the bitmask system works. there was a sufficient description on how to sum up the 9 booleans into one number, but I had to figure out how to turn that number back into booleans myself. I took a bit of a guess and it worked. if this function has a name, then it'd be good to say that name so people have an idea of what to Google, but I can't seem to find that name. so here I've done my best to describe what the function needs to do while keeping my description short, but I don't know if this is the right place to put it or if my description makes enough sense. I needed this to make my own logic for autotiles, as I'm dealing with destructable terrain and the default behavior in update_bitmask_area() didn't do everything I wanted.